Created separate package for AOP concepts question
