+++ src/main/java/spring/certification/ioc/q024/SpringAOP.java
@@ -1,6 +1,6 @@
 package spring.certification.ioc.q024;
 
-import spring.certification.aop.Q001concepts;
+import spring.certification.aop.q001.AOPConcepts;
 import spring.certification.aop.Q003proxy;
 import spring.certification.aop.Q004limitations;
 
@@ -16,7 +16,7 @@ import spring.certification.aop.Q004limitations;
  * AOP implementation details as well as different types of proxy objects are discussed in the {@link Q003proxy proxy}
  * question of the AOP certification section.<br>
  * Limitations are discussed in the {@link Q004limitations limitations} question of the AOP certification section.<br>
- * Advantages of using proxies are discussed in the {@link Q001concepts consepts} question of the AOP certification section.<br>
+ * Advantages of using proxies are discussed in the {@link AOPConcepts consepts} question of the AOP certification section.<br>
  *
  * @author Valentine Shemyako
  * @since November 30, 2018
